requirements.txt
black==23.7.0
Django==4.2
djangorestframework==3.14.0
ipython==8.14.0
isort==5.12.0
pytest==7.4.0
pytest-django==4.5.2
